fix(shared): move context window catalog to @cat-cafe/shared (#1343)
Why: draftCannotResolveOrApplyContextWindow hardcoded Google clientId as
always-incompatible, but resolveContextCapacity would resolve cataloged
models (gemini-2.5-pro etc.) via the model catalog. The invariant is:
Auto unresolved ⇔ no carrier report AND no model catalog entry.

What:
- Extract CONTEXT_WINDOW_SIZES + getContextWindowFallback to
  @cat-cafe/shared/context-window-catalog (pure static data, zero
  runtime deps — same pattern as avatar-limits, codex-speed)
- API context-window-sizes re-exports from shared and keeps
  runtime-only helpers (KNOWN_MIN, resolveContextWindow)
- Web draft heuristic now checks model catalog before warning:
  antigravity/google + cataloged model → no warning
- Test updated: gemini-2.5-pro draft no longer shows false warning;
  unknown models still warn correctly

[宪宪/claude-opus-4-6🐾]
